Adding initial github actions (#6)
Adds 

- A `Test` workflow to build, test and calculate coverage for the module
- A `Deploy` workflow for building a ghpages site

This PR does not attempt to resolve all package issues, only to get the
infrastructure set up as a starting point. It appears that at the very
least we have to update the development manifest to test against the
latest julia release, or better yet remove the manifest since I think
that's only considered a best practice for things like reproducible
analysis.